# apps/api/openapi.yaml
openapi: 3.0.3
info:
  title: RISE API
  version: 1.0.0
  description: RISE — Autonomous Incident Remediation System API Specification
servers:
  - url: https://api.rise.internal/api/v1
    description: Internal API Gateway
security:
  - BearerAuth: []
paths:
...
  /incidents/{incident_id}:
    get:
      summary: Full incident detail
      operationId: getIncident
      parameters:
        - name: incident_id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Expanded incident details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IncidentDetailEnvelope'
        '401':
          $ref: '#/components/responses/401Unauthorized'
        '404':
          $ref: '#/components/responses/404NotFound'
        '422':
          $ref: '#/components/responses/422ValidationError'

    patch:
      summary: Update severity or status manually
      operationId: patchIncident
      parameters:
        - name: incident_id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/IncidentUpdateRequest'
      responses:
        '200':
          description: Incident updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IncidentEnvelope'
        '401':
          $ref: '#/components/responses/401Unauthorized'
        '404':
          $ref: '#/components/responses/404NotFound'
        '422':
          $ref: '#/components/responses/422ValidationError'
...
  /incidents/{incident_id}/decision:
    get:
      summary: Decision and action plan recommendation
      operationId: getDecision
      parameters:
        - name: incident_id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Decision and action plan
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DecisionEnvelope'
        '401':
          $ref: '#/components/responses/401Unauthorized'
        '404':
          $ref: '#/components/responses/404NotFound'
        '422':
          $ref: '#/components/responses/422ValidationError'
  /incidents/{incident_id}/actions/{action_id}/approve:
    post:
      summary: Approve a pending high-risk action
      operationId: approveAction
      parameters:
        - name: incident_id
          in: path
          required: true
          schema:
            type: string
        - name: action_id
          in: path
          required: true
          schema:
            type: string
        - name: Idempotency-Key
          in: header
          required: true
          schema:
            type: string
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ActionApproveRequest'
      responses:
        '200':
          description: Action approved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ActionApproveEnvelope'
        '401':
          $ref: '#/components/responses/401Unauthorized'
        '404':
          $ref: '#/components/responses/404NotFound'
        '409':
          $ref: '#/components/responses/409ActionPlanChanged'
        '422':
          $ref: '#/components/responses/422ValidationError'
  /incidents/{incident_id}/actions/{action_id}/reject:
    post:
      summary: Reject a pending action
      operationId: rejectAction
      parameters:
        - name: incident_id
          in: path
          required: true
          schema:
            type: string
        - name: action_id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ActionRejectRequest'
      responses:
        '200':
          description: Action rejected
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ActionRejectEnvelope'
        '401':
          $ref: '#/components/responses/401Unauthorized'
        '404':
          $ref: '#/components/responses/404NotFound'
        '422':
          $ref: '#/components/responses/422ValidationError'
  /incidents/{incident_id}/actions/{action_id}/modify:
    post:
      summary: Submit modified action plan for re-evaluation
      operationId: modifyAction
      parameters:
        - name: incident_id
          in: path
          required: true
          schema:
            type: string
        - name: action_id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ActionModifyRequest'
      responses:
        '200':
          description: Action plan re-evaluated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ActionModifyEnvelope'
        '401':
          $ref: '#/components/responses/401Unauthorized'
        '404':
          $ref: '#/components/responses/404NotFound'
        '422':
          $ref: '#/components/responses/422ValidationError'
  /incidents/{incident_id}/actions:
    get:
      summary: List all remediation actions for an incident
      operationId: listActions
      parameters:
        - name: incident_id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Remediation actions list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RemediationActionListEnvelope'
        '401':
          $ref: '#/components/responses/401Unauthorized'
        '404':
          $ref: '#/components/responses/404NotFound'
        '422':
          $ref: '#/components/responses/422ValidationError'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    Meta:
      type: object
      required:
        - request_id
        - timestamp
      properties:
        request_id:
          type: string
        timestamp:
          type: string
          format: date-time
        next_cursor:
          type: string
          nullable: true
    ErrorDetail:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: string
          enum:
            - VALIDATION_ERROR
            - RISK_POLICY_VIOLATION
            - APPROVAL_EXPIRED
            - ACTION_PLAN_CHANGED
            - RESOURCE_LOCKED
            - INTEGRATION_UNAVAILABLE
            - RATE_LIMITED
            - UNAUTHORIZED
            - NOT_FOUND
        message:
          type: string
        details:
          type: object
    ErrorEnvelope:
      type: object
      required:
        - data
        - meta
        - error
      properties:
        data:
          type: object
          nullable: true
          default: null
        meta:
          $ref: '#/components/schemas/Meta'
        error:
          $ref: '#/components/schemas/ErrorDetail'
...
    Incident:
      type: object
      required: [id, title, description, severity, status, created_at]
      properties:
        id:
          type: string
        title:
          type: string
        description:
          type: string
        severity:
          type: string
          enum: [SEV1, SEV2, SEV3, SEV4]
        status:
          type: string
          enum: [open, investigating, awaiting_approval, resolved, closed]
        affected_service:
          type: string
          nullable: true
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
          nullable: true
        resolution_note:
          type: string
          nullable: true
    IncidentDetail:
      allOf:
        - $ref: '#/components/schemas/Incident'
        - type: object
          properties:
            timeline:
              type: array
              items:
                type: object
            root_cause:
              type: object
              nullable: true
            impact:
              type: object
              nullable: true
            actions:
              type: array
              items:
                type: object
            approvals:
              type: array
              items:
                type: object
...
    IncidentUpdateRequest:
      type: object
      required: [status, resolution_note]
      properties:
        status:
          type: string
        resolution_note:
          type: string
...
    IncidentEnvelope:
      type: object
      required: [data, meta, error]
      properties:
        data:
          $ref: '#/components/schemas/Incident'
        meta:
          $ref: '#/components/schemas/Meta'
        error:
          type: object
          nullable: true
    IncidentDetailEnvelope:
      type: object
      required: [data, meta, error]
      properties:
        data:
          $ref: '#/components/schemas/IncidentDetail'
        meta:
          $ref: '#/components/schemas/Meta'
        error:
          type: object
          nullable: true
...
    ActionPlan:
      type: object
      required: [id, description, steps]
      properties:
        id:
          type: string
        description:
          type: string
        steps:
          type: array
          items:
            type: string
    Decision:
      type: object
      required: [risk_tier, confidence, recommended_action, requires_approval]
      properties:
        risk_tier:
          type: string
        confidence:
          type: number
        recommended_action:
          $ref: '#/components/schemas/ActionPlan'
        requires_approval:
          type: boolean
    DecisionEnvelope:
      type: object
      required: [data, meta, error]
      properties:
        data:
          $ref: '#/components/schemas/Decision'
        meta:
          $ref: '#/components/schemas/Meta'
        error:
          type: object
          nullable: true
    ActionApproveRequest:
      type: object
      properties:
        note:
          type: string
          nullable: true
    ActionApproveData:
      type: object
      required: [status, execution_status]
      properties:
        status:
          type: string
        execution_status:
          type: string
    ActionApproveEnvelope:
      type: object
      required: [data, meta, error]
      properties:
        data:
          $ref: '#/components/schemas/ActionApproveData'
        meta:
          $ref: '#/components/schemas/Meta'
        error:
          type: object
          nullable: true
    ActionRejectRequest:
      type: object
      required: [reason]
      properties:
        reason:
          type: string
    ActionRejectData:
      type: object
      required: [status]
      properties:
        status:
          type: string
    ActionRejectEnvelope:
      type: object
      required: [data, meta, error]
      properties:
        data:
          $ref: '#/components/schemas/ActionRejectData'
        meta:
          $ref: '#/components/schemas/Meta'
        error:
          type: object
          nullable: true
    ActionModifyRequest:
      type: object
      required: [modified_plan]
      properties:
        modified_plan:
          $ref: '#/components/schemas/ActionPlan'
    ActionModifyData:
      type: object
      required: [status, new_risk_tier]
      properties:
        status:
          type: string
        new_risk_tier:
          type: string
    ActionModifyEnvelope:
      type: object
      required: [data, meta, error]
      properties:
        data:
          $ref: '#/components/schemas/ActionModifyData'
        meta:
          $ref: '#/components/schemas/Meta'
        error:
          type: object
          nullable: true
    RemediationAction:
      type: object
      required: [id, incident_id, name, risk_tier, status]
      properties:
        id:
          type: string
        incident_id:
          type: string
        name:
          type: string
        risk_tier:
          type: string
        status:
          type: string
    RemediationActionListEnvelope:
      type: object
      required: [data, meta, error]
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/RemediationAction'
        meta:
          $ref: '#/components/schemas/Meta'
        error:
          type: object
          nullable: true
...
  responses:
    401Unauthorized:
      description: Unauthorized - missing or invalid token
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
          example:
            data: null
            meta:
              request_id: "00000000-0000-0000-0000-000000000000"
              timestamp: "2026-08-01T00:00:00Z"
            error:
              code: "UNAUTHORIZED"
              message: "Missing or invalid authorization token"
              details: {}
...
    404NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
          example:
            data: null
            meta:
              request_id: "00000000-0000-0000-0000-000000000000"
              timestamp: "2026-08-01T00:00:00Z"
            error:
              code: "NOT_FOUND"
              message: "Requested resource was not found"
              details: {}
    409ActionPlanChanged:
      description: Conflict - action plan hash changed or approval expired
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
          example:
            data: null
            meta:
              request_id: "00000000-0000-0000-0000-000000000000"
              timestamp: "2026-08-01T00:00:00Z"
            error:
              code: "ACTION_PLAN_CHANGED"
              message: "Action plan hash changed since approval was requested"
              details: {}
    422ValidationError:
      description: Unprocessable Entity - validation error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
          example:
            data: null
            meta:
              request_id: "00000000-0000-0000-0000-000000000000"
              timestamp: "2026-08-01T00:00:00Z"
            error:
              code: "VALIDATION_ERROR"
              message: "Request body failed schema validation"
              details: {}
...